test fixed

+++ src/app/config/TEST.py
@@ -12,5 +12,5 @@ class TEST(BaseSettings):
 
     questions_path: str = Field("load_test/questions.json")
     repo_id: int = Field(0)
-    think_min: int = Field(0.2)
-    think_max: int = Field(0.6)
+    think_min: float = Field(0.2)
+    think_max: float = Field(0.6)
